+++ packages/flutter/lib/widgets.dart
@@ -46,10 +46,11 @@ export 'src/widgets/performance_overlay.dart';
 export 'src/widgets/placeholder.dart';
 export 'src/widgets/raw_keyboard_listener.dart';
 export 'src/widgets/routes.dart';
-export 'src/widgets/scroll_absolute.dart';
 export 'src/widgets/scroll_behavior.dart';
 export 'src/widgets/scroll_configuration.dart';
 export 'src/widgets/scroll_notification.dart';
+export 'src/widgets/scroll_physics.dart';
+export 'src/widgets/scroll_position.dart';
 export 'src/widgets/scroll_simulation.dart';
 export 'src/widgets/scroll_view.dart';
 export 'src/widgets/scrollable.dart';
