+++ packages/flutter/test/widgets/list_test.dart
@@ -72,7 +72,7 @@ void main() {
     log.clear();
 
     Scrollable2State state = tester.state(find.byType(Scrollable2));
-    AbsoluteScrollPosition position = state.position;
+    ScrollPosition position = state.position;
     position.jumpTo(2025.0);
 
     expect(log, isEmpty);
